+++ adept/lpse2d/core/integrator.py
@@ -2,17 +2,8 @@ from typing import Dict, List
 
 from jax import numpy as jnp
 import equinox as eqx
-import diffrax
 
 from adept.lpse2d.core import epw
-
-
-class Stepper(diffrax.Euler):
-    def step(self, terms, t0, t1, y0, args, solver_state, made_jump):
-        del solver_state, made_jump
-        y1 = terms.vf(t0, y0, args)
-        dense_info = dict(y0=y0, y1=y1)
-        return y1, None, dense_info, None, diffrax.RESULTS.successful
 
 
 class VectorField(eqx.Module):
